Separate JavaScript in 4 files for easy managing and to avoid conflicts between pages. No updates yet.
I decided to avoid dumping all js in one place as the file grows bigger and bigger and it became hard to manage it.

Split it in 4: shared.js (linked to all pages) and separate for all 3 html files.

Script.js is now redundant, left untouched (yet) only for comparison.